Non-working semantic analysis implementation.
This commit "finalizes" working the name resolution and type-checking
processes into the compiler's overall structure and updates the build
system to support the now non-parser-centric output. However, the name
resolution process seems to loop indefinitely on `ht_get()`. Will need
to troubleshoot or possibly find a different hash table implementation.

Also modifies the `scoping.bm` example. There's no obvious way to work
comma-separated print expressions into the print statement the way the
book shows, or certainly not without considerable reworking. Because this
functionality, as far as I know, wasn't shown in the description of
B-Minor's grammar, I'm treating it as an extra feature to possibly
add in the future.